[REFACTOR] #6 - Entity layer 분리
