+++ src/main/kotlin/com/example/smartshopping/entity/PurchaseRecord.kt
@@ -1,6 +1,6 @@
-package com.example.smartshopping.domain.purchase
+package com.example.smartshopping.entity
 
-import com.example.smartshopping.domain.jpa.BaseEntity
+import com.example.smartshopping.entity.base.BaseEntity
 import javax.persistence.Entity
 
 @Entity(name = "purchase")
