+++ src/main/kotlin/com/example/smartshopping/entity/User.kt
@@ -1,6 +1,6 @@
-package com.example.smartshopping.domain.user
+package com.example.smartshopping.entity
 
-import com.example.smartshopping.domain.jpa.BaseEntity
+import com.example.smartshopping.entity.base.BaseEntity
 import javax.persistence.Entity
 
 @Entity(name = "user")
